+++ packages/video_player/video_player_android/android/src/test/java/io/flutter/plugins/videoplayer/ExoPlayerEventListenerTest.java
@@ -17,7 +17,6 @@ import androidx.media3.common.PlaybackException;
 import androidx.media3.common.Player;
 import androidx.media3.common.VideoSize;
 import androidx.media3.exoplayer.ExoPlayer;
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -45,7 +44,8 @@ public final class ExoPlayerEventListenerTest {
   @Test
   @Config(maxSdk = 21)
   public void onPlaybackStateChangedReadySendInitialized_belowAndroid21() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     VideoSize size = new VideoSize(800, 400, 0, 0);
     when(mockExoPlayer.getVideoSize()).thenReturn(size);
     when(mockExoPlayer.getDuration()).thenReturn(10L);
@@ -56,8 +56,10 @@ public final class ExoPlayerEventListenerTest {
 
   @Test
   @Config(minSdk = 22)
-  public void onPlaybackStateChangedReadySendInitialized_whenSurfaceProducerHandlesCropAndRotation() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+  public void
+      onPlaybackStateChangedReadySendInitialized_whenSurfaceProducerHandlesCropAndRotation() {
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     VideoSize size = new VideoSize(800, 400, 0, 0);
     when(mockExoPlayer.getVideoSize()).thenReturn(size);
     when(mockExoPlayer.getDuration()).thenReturn(10L);
@@ -70,7 +72,8 @@ public final class ExoPlayerEventListenerTest {
   @Config(minSdk = 22)
   public void
       onPlaybackStateChangedReadySendInitializedWithRotationCorrectionAndWidthAndHeightSwap_whenSurfaceProducerDoesNotHandleCropAndRotation() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, false);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, false);
     VideoSize size = new VideoSize(800, 400, 0, 0);
     int rotationCorrection = 90;
     Format videoFormat = new Format.Builder().setRotationDegrees(rotationCorrection).build();
@@ -87,7 +90,8 @@ public final class ExoPlayerEventListenerTest {
   @Config(maxSdk = 21)
   public void
       onPlaybackStateChangedReadyInPortraitMode90DegreesSwapWidthAndHeight_belowAndroid21() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     VideoSize size = new VideoSize(800, 400, 90, 0);
     when(mockExoPlayer.getVideoSize()).thenReturn(size);
     when(mockExoPlayer.getDuration()).thenReturn(10L);
@@ -100,7 +104,8 @@ public final class ExoPlayerEventListenerTest {
   @Config(minSdk = 22)
   public void
       onPlaybackStateChangedReadyInPortraitMode90DegreesDoesNotSwapWidthAndHeight_whenSurfaceProducerHandlesCropAndRotation() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     VideoSize size = new VideoSize(800, 400, 90, 0);
 
     when(mockExoPlayer.getVideoSize()).thenReturn(size);
@@ -114,7 +119,8 @@ public final class ExoPlayerEventListenerTest {
   @Config(minSdk = 22)
   public void
       onPlaybackStateChangedReadyInPortraitMode90DegreesSwapWidthAndHeight_whenSurfaceProducerDoesNotHandleCropAndRotation() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, false);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, false);
     VideoSize size = new VideoSize(800, 400, 0, 0);
     int rotationCorrection = 90;
     Format videoFormat = new Format.Builder().setRotationDegrees(rotationCorrection).build();
@@ -131,7 +137,8 @@ public final class ExoPlayerEventListenerTest {
   @Config(maxSdk = 21)
   public void
       onPlaybackStateChangedReadyInPortraitMode270DegreesSwapWidthAndHeight_belowAndroid21() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     VideoSize size = new VideoSize(800, 400, 270, 0);
     when(mockExoPlayer.getVideoSize()).thenReturn(size);
     when(mockExoPlayer.getDuration()).thenReturn(10L);
@@ -144,7 +151,8 @@ public final class ExoPlayerEventListenerTest {
   @Config(minSdk = 22)
   public void
       onPlaybackStateChangedReadyInPortraitMode270DegreesDoesNotSwapWidthAndHeight_whenSurfaceProducerHandlesCropAndRotation() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     VideoSize size = new VideoSize(800, 400, 270, 0);
     when(mockExoPlayer.getVideoSize()).thenReturn(size);
     when(mockExoPlayer.getDuration()).thenReturn(10L);
@@ -157,7 +165,8 @@ public final class ExoPlayerEventListenerTest {
   @Config(minSdk = 22)
   public void
       onPlaybackStateChangedReadyInPortraitMode270DegreesSwapWidthAndHeight_whenSurfaceProducerDoesNotHandleCropAndRotation() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, false);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, false);
     VideoSize size = new VideoSize(800, 400, 0, 0);
     int rotationCorrection = 270;
     Format videoFormat = new Format.Builder().setRotationDegrees(rotationCorrection).build();
@@ -173,7 +182,8 @@ public final class ExoPlayerEventListenerTest {
   @Test
   @Config(maxSdk = 21)
   public void onPlaybackStateChangedReadyFlipped180DegreesInformEventHandler_belowAndroid21() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     VideoSize size = new VideoSize(800, 400, 180, 0);
     when(mockExoPlayer.getVideoSize()).thenReturn(size);
     when(mockExoPlayer.getDuration()).thenReturn(10L);
@@ -184,7 +194,8 @@ public final class ExoPlayerEventListenerTest {
 
   @Test
   public void onPlaybackStateChangedBufferingSendsBufferingStartAndUpdates() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     when(mockExoPlayer.getBufferedPosition()).thenReturn(10L);
     eventListener.onPlaybackStateChanged(Player.STATE_BUFFERING);
 
@@ -199,7 +210,8 @@ public final class ExoPlayerEventListenerTest {
 
   @Test
   public void onPlaybackStateChangedEndedSendsOnCompleted() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     eventListener.onPlaybackStateChanged(Player.STATE_ENDED);
 
     verify(mockCallbacks).onCompleted();
@@ -208,7 +220,8 @@ public final class ExoPlayerEventListenerTest {
 
   @Test
   public void onPlaybackStateChangedEndedAfterBufferingSendsBufferingEndAndOnCompleted() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     when(mockExoPlayer.getBufferedPosition()).thenReturn(10L);
     eventListener.onPlaybackStateChanged(Player.STATE_BUFFERING);
     verify(mockCallbacks).onBufferingStart();
@@ -223,7 +236,8 @@ public final class ExoPlayerEventListenerTest {
 
   @Test
   public void onPlaybackStateChangedIdleDoNothing() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     eventListener.onPlaybackStateChanged(Player.STATE_IDLE);
 
     verifyNoInteractions(mockCallbacks);
@@ -231,7 +245,8 @@ public final class ExoPlayerEventListenerTest {
 
   @Test
   public void onPlaybackStateChangedIdleAfterBufferingSendsBufferingEnd() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     when(mockExoPlayer.getBufferedPosition()).thenReturn(10L);
     eventListener.onPlaybackStateChanged(Player.STATE_BUFFERING);
     verify(mockCallbacks).onBufferingStart();
@@ -245,7 +260,8 @@ public final class ExoPlayerEventListenerTest {
 
   @Test
   public void onErrorVideoErrorWhenBufferingInProgressAlsoEndBuffering() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     when(mockExoPlayer.getBufferedPosition()).thenReturn(10L);
     eventListener.onPlaybackStateChanged(Player.STATE_BUFFERING);
     verify(mockCallbacks).onBufferingStart();
@@ -259,7 +275,8 @@ public final class ExoPlayerEventListenerTest {
 
   @Test
   public void onErrorBehindLiveWindowSeekToDefaultAndPrepare() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     eventListener.onPlayerError(
         new PlaybackException("SORT_OF_OK", null, PlaybackException.ERROR_CODE_BEHIND_LIVE_WINDOW));
 
@@ -270,7 +287,8 @@ public final class ExoPlayerEventListenerTest {
 
   @Test
   public void onIsPlayingChangedToggled() {
-    ExoPlayerEventListener eventListener = new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
+    ExoPlayerEventListener eventListener =
+        new ExoPlayerEventListener(mockExoPlayer, mockCallbacks, true);
     eventListener.onIsPlayingChanged(true);
     verify(mockCallbacks).onIsPlayingStateUpdate(true);